+++ src/components/modal_component/ModalComponent.h
@@ -35,7 +35,7 @@ namespace Project::Components {
     void build(Project::Utilities::LuaStateWrapper& luaStateWrapper, const std::string& tableName) override;
     void applyStyle() override;
 
-    void setEntityPosition(int x, int y) override;
+    void setEntityPosition(float x, float y) override;
     void setSize(int w, int h);
 
     void setEntityReference(Project::Entities::Entity* entity) { owner = entity; }
